Adds a filter before returning content
This commit adds a filter that cann be used to wrap or alter the content of the
included page.

It introduces the filter ```insert_pages_wrap_content``` that will call a
callback with three parameters:

 1. The content
 2. The array of Posts that matched the criteria (should be 0 or 1
 entry)
 3. The array of parameters passed from the shortcode.

That way you can add a filter to f.i. add the featured image of the post
you are includng before the inserted content by adding the following to
your themes function.php:

    add_filter(
        'insert_pages_wrap_content',
        function($content, $posts,$parameters){
            return get_the_post_thumbnail($posts[0]->ID, 'large') .
                $content;
        },
        9,
        3
    );