Don't use query_posts unless using custom templates
The WordPress Codex cautions against using query_posts since it is
inefficient and legacy core code. See:
https://codex.wordpress.org/Function_Reference/query_posts
We only use it to allow the use of template tags in custom templates
(i.e., we replace the main loop with the inserted page’s loop so all
template tags work as-is, and then restore the main loop afterwards).
This commit factors out query_posts in favor of get_page_by_path() or
get_post() for all built-in displays, and only uses query_posts for
custom template displays.