Restrict custom template paths to theme directory
WordPress’s locate_template() function does not enforce that template files must be in the theme directory, so a directory traversal attack could be used to expose information about the running server. This fix restricts template files to the active theme directory or wp-includes/theme-compat by not allowing any directory traversal (e.g., ../) tokens in the template path. (Note: realpath() will calcuate the absolute path to the file, so if it’s different than the user-supplied path, that means the user-supplied path has directory traversal tokens in it.)